Add datatype to ingredient (#154)
* Add dtype to ingredient

Fix caching issue w dtype generation and lark grammar

* Add dtype to ingredient

Fix caching issue w dtype generation and lark grammar

* Require num for metrics

* Reorganize utils

* WIP

* Create dtype utils

* Track dtype

* Rename dtype to datatype

* Remove enum

* auto-format Python code

* Fixes
