+++ recipe/schemas/parsed_constructors.py
@@ -59,7 +59,7 @@ def _convert_bucket_to_field(field, bucket, buckets_default_label, builder):
 
 
 def create_ingredient_from_parsed(ingr_dict, builder, debug=False):
-    """ Create an ingredient from config version 2 object . """
+    """Create an ingredient from config version 2 object ."""
     kind = ingr_dict.pop("kind", "metric")
     IngredientClass = ingredient_class_for_name(kind.title())
     if IngredientClass is None:
@@ -72,9 +72,19 @@ def create_ingredient_from_parsed(ingr_dict, builder, debug=False):
             if kind == "metric":
                 fld_defn = ingr_dict.pop("field", None)
                 # SQLAlchemy ingredient with required aggregation
-                expr = builder.parse(fld_defn, enforce_aggregation=True, debug=debug)
+                expr, datatype = builder.parse(
+                    fld_defn, enforce_aggregation=True, debug=debug
+                )
                 # Save the data type in the ingredient
-                ingr_dict["_data_type"] = builder.last_datatype
+                ingr_dict["datatype"] = datatype
+                if datatype != "num":
+                    error = {
+                        "type": "Can not parse field",
+                        "extra": {
+                            "details": "A string can not be aggregated",
+                        },
+                    }
+                    return InvalidIngredient(error=error)
                 args = [expr]
             else:
                 fld_defn = ingr_dict.pop("field", None)
@@ -89,38 +99,57 @@ def create_ingredient_from_parsed(ingr_dict, builder, debug=False):
                     ingr_dict["extra_fields"].append(
                         {"name": "order_by_expression", "field": order_by_fld}
                     )
-                expr = builder.parse(fld_defn, forbid_aggregation=True, debug=debug)
+                expr, datatype = builder.parse(
+                    fld_defn, forbid_aggregation=True, debug=debug
+                )
                 # Save the data type in the ingredient
-                ingr_dict["_data_type"] = builder.last_datatype
+                ingr_dict["datatype"] = datatype
                 args = [expr]
+
                 # Convert extra fields to sqlalchemy expressions and add them directly to
-                # the kwargs
+                # the kwargs, saving datatypes
+                datatype_by_role = {"value": datatype}
                 for extra in ingr_dict.pop("extra_fields", []):
-                    ingr_dict[extra.get("name")] = builder.parse(
+                    raw_role = extra.get("name")
+                    if raw_role.endswith("_expression"):
+                        # Remove _expression to get the role
+                        role = raw_role[:-11]
+                    else:
+                        role = raw_role
+
+                    expr, datatype = builder.parse(
                         extra.get("field"), forbid_aggregation=True, debug=debug
                     )
+                    datatype_by_role[role] = datatype
+                    ingr_dict[raw_role] = expr
+                ingr_dict["datatype_by_role"] = datatype_by_role
 
             parsed_quickselects = []
             for qs in ingr_dict.pop("quickselects", []):
                 condition_defn = qs.get("condition")
+                expr, _ = builder.parse(
+                    condition_defn, forbid_aggregation=True, debug=debug
+                )
                 parsed_quickselects.append(
                     {
                         "name": qs["name"],
-                        "condition": builder.parse(
-                            condition_defn, forbid_aggregation=True, debug=debug
-                        ),
+                        "condition": expr,
                     }
                 )
             ingr_dict["quickselects"] = parsed_quickselects
 
         elif kind == "filter":
             condition_defn = ingr_dict.get("condition")
-            args = [builder.parse(condition_defn, forbid_aggregation=True, debug=debug)]
+            expr, _ = builder.parse(
+                condition_defn, forbid_aggregation=True, debug=debug
+            )
+            args = [expr]
         elif kind == "having":
             condition_defn = ingr_dict.get("condition")
-            args = [
-                builder.parse(condition_defn, forbid_aggregation=False, debug=debug)
-            ]
+            expr, _ = builder.parse(
+                condition_defn, forbid_aggregation=False, debug=debug
+            )
+            args = [expr]
 
     except (GrammarError, LarkError) as e:
         error_msg = str(e)
